Add files via upload
# DailyBinge-Your Personal Chef!
In today's fast-paced world, people are busy with their jobs and many of them don't know how to cook and feel that investing time and strength in learning to cook will only waste their precious hours. 
People look for fast ways to learn cooking but if not found, ultimately will rely on online food ordering. 
Particularly, in the last two years of COVID-19, 73% of people enjoyed lockdown cooking,creating different recipes by browsing on various platforms. 
Our app is a new concise version of a 'learn to cook' app which will help beginners as well as mediocre to learn cooking in step by step descriptive and engaging manner.
The user will be able to learn cooking in a guided and efficient way with minimal cooking faults.